+++ src/main/java/net/weeek/denisbaranov/utils/Waiters.java
@@ -1,9 +1,12 @@
-package net.weeek.denisbaranov;
+package net.weeek.denisbaranov.utils;
 
 public class Waiters {
+
+    public static final int DEFAULT_WAIT_SECONDS = 4;
+
     public static void waitFor(int seconds) {
         try {
-            Thread.sleep(seconds*1000);
+            Thread.sleep(seconds * 1000);
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
