+++ src/main/java/net/weeek/denisbaranov/xpathes/HomePageXPath.java
@@ -1,4 +1,4 @@
-package net.weeek.denisbaranov;
+package net.weeek.denisbaranov.xpathes;
 
 public class HomePageXPath {
     public static final String BUTTON_LOGIN_XPATH = "//a[@class='header__button s button-component white-fill ux-radial']";
